+++ src/libgenesis13/loading/ImportField.cpp
@@ -57,7 +57,7 @@ bool ImportField::init(int rank, int size, map<string, string>* arg,
     time->getPosition(&s);
     int idx=-1;
     Field* field;
-    for (int i=0; i<fieldin->size(); i++) {
+    for (size_t i=0; i<fieldin->size(); i++) {
         if (fieldin->at(i)->harm==harm) {
             field=fieldin->at(i);
             idx=i;
